libcamera: Infrastructure to ask for "memory" cameras
"Memory" cameras are used to make it easier to process raw camera
images that are stored in memory buffers. There may be no actual
camera attached to the system at all.

Memory cameras are not created when the system starts. Depending on
the pipeline handler, it may be possible to open many memory cameras
simultaneously, or none at all. So any time an application requests a
memory camera, we check with the pipeline handler whether it can
provide one.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>