libcamera: pipeline: rpi: Allow creation of the first "memory" camera
The Raspberry Pi PiSP pipeline handler is updated to indicate that it
supports the creation of "memory" cameras, and extended actually to do
so.

Memory cameras are created on demand when applications request them.

Subsequent commits will support the correct handling of camera
configurations with "input" streams, that will require the use of
these memory cameras.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>