libcamera: sensor: Add CameraSensorMemory class
Representation of a "camera" that actually takes its input from
a memory buffer.

libcamera is, unsurprisingly, very dependent on having a camera
connected to the system. But sometimes we may want to process raw
camera images from elsewhere, and we may not have a camera connected
to the system at all.

In such cases, the path of a "memory buffer" through the code is eased
considerably by introducing the CameraSensorMemory class, which allows
the memory buffer to behave, to an extent at least, like a real
camera.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>