pipeline: rpi: Allow processing of raw input streams
This is the first version that will successfully perform "Bayer
reprocessing" by taking a raw input stream, feeding it to the PiSP
Back End, and producing a correct output image.

Although the CameraSensorMemory and CamHelperDefault classes help with
this, there is a horrifying about of "whack-a-mole", just to avoid
doing things - mostly involving the Camera Front End block - that will
just crash.

Obviously this will all need more careful thought.

This commit only updates the pipeline handler, a subsequent commit
will fix up similar issues in the IPA.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>