ipa: rpi: Minor workarounds for memory cameras
We are missing stuff like embedded data, so need to avoid falling
over when it's not there. Also stuff like lens controls need to
be avoided.

We also need to take care that the system doesn't start up with
zero digital gains, as that will give us a zero output!

You get the idea...

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>