ipa: rpi: Support memory cameras
Add support for "memory cameras" (for Bayer reprocessing).

Mostly very minor changes, but the way we fill in certain bits of
metadata before runningthe IPAs changes significantly (see
prepareIspFillMetadata).

For example, regular sensors get device status information from the
the sensor (or possibly from values we wrote to the sensor a few
frames back), whereas "memory cameras" assume that the exposure/gain
values passed in with the request controls will be the ones for
setting up the tuning algorithms.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>