pipeline: rpi: Support memory cameras
Mostly small changes to support memory cameras (for Bayer reprocessing)
in the pipeline handler.

Firstly, we make a "CameraSensorMemory" as the sensor as soon as we
have the raw input buffer configuration, which helps the rest of the
code to work with fewer changes.

The PiSP platformConfigure method is refactored to split the Cfe and
Isp parts into separate functions. Memory cameras then omit
platformConfigureCfe but run platformConfigureIsp as before.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>